+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## v0.26 -- ?
 
 - Added `F12` to show original screen until a key is pressed.
+- Added `-f`, `--find`, `-i`, `--ignore-case`, `-i-`, `--exact-case`, `-r`, and `--regex` command line options for searching.
 - Added custom color for file content.
 - Added horizontal scroll markers when scrolling input text horizontally.
 - Added mouse input for input text.